% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_control.R
\name{cv_control}
\alias{cv_control}
\title{Create Cross-Validation Settings}
\usage{
cv_control(
  method = "repeatedcv",
  number = 5,
  repeats = 1,
  classProbs = TRUE,
  summaryFunction = caret::twoClassSummary
)
}
\arguments{
\item{method}{Type of cross validation}

\item{number}{Amount of folds to split the training data into}

\item{repeats}{Amount of times repeating the cross validation process}

\item{classProbs}{Logical value indicating whether class probabilities should be computed}

\item{summaryFunction}{A function used to calculate model performance metrics}
}
\value{
A \code{trainControl} object that can be passed to \code{caret::train()}.
}
\description{
Create Cross-Validation Settings
}
\examples{
cv_control()
}
