% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgrid.R
\name{kgrid}
\alias{kgrid}
\title{Create a K Grid to test K values}
\usage{
kgrid(min_k, max_k, step)
}
\arguments{
\item{min_k}{numeric value for minimum k value to test}

\item{max_k}{numeric value for the maximum k value to test}

\item{step}{numeric value for the size of each jump between k values}
}
\value{
a data frame with one column, \code{k}, designed to hold tuning values
}
\description{
Create a K Grid to test K values
}
\examples{
kgrid(1, 11, 2)
}
